Implemented add music album to items collection
