+++ main.rb
@@ -94,84 +94,7 @@ end
 
 # adding game   [END]........................................................................................
 
-# music_album   [START]>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
-
-def save_genres(genres)
-  genre_data = genres.map do |genre|
-    {
-      'id' => genre.id,
-      'genre_name' => genre.name || 'No Genre'
-    }
-  end
-
-  File.write('genre.json', JSON.pretty_generate(genre_data))
-rescue JSON::GeneratorError => e
-  puts "Error generating JSON data for 'genre.json': #{e.message}"
-end
-
-# load and save data to json file [START]......
-def load_music_albums(genres)
-  return [] unless File.exist?('music.json')
-
-  json_data = File.read('music.json')
-  music_albums_data = JSON.parse(json_data)
-
-  music_albums_data.map do |album_data|
-    load_music_album(album_data, genres)
-  end
-rescue JSON::ParserError => e
-  puts "Error parsing 'music.json': #{e.message}"
-  []
-end
-
-def load_music_album(album_data, genres)
-  album = album_data['album']
-  id = album_data['id']
-  album_name = album['album_name']
-  can_be_archived = album['can_be_archived']
-  on_spotify = album['on_spotify']
-  genre_name = extract_genre_name(album['genre'])
-
-  genre = find_or_create_genre(genres, genre_name)
-
-  puts "Loaded music album: ID=#{id}, Name=#{album_name},
-   Can Be Archived=#{can_be_archived}, On Spotify=#{on_spotify}, Genre=#{genre_name}"
-
-  MusicAlbum.new(id, album_name, on_spotify, can_be_archived, genre)
-end
-
-def extract_genre_name(genre_data)
-  genre_data.is_a?(Hash) && genre_data.key?('name') ? genre_data['name'] : nil
-end
-
-def find_or_create_genre(genres, genre_name)
-  genre = genres.find { |g| g.name == genre_name }
-  unless genre
-    genre = Genre.new(genres.size + 1, genre_name)
-    genres << genre
-  end
-  genre
-end
-
-def load_genres
-  if File.exist?('genre.json')
-    json_data = File.read('genre.json')
-    return [] if json_data.strip.empty?
-
-    JSON.parse(json_data).map do |genre_data|
-      id = genre_data['id']
-      name = genre_data['genre_name']
-      Genre.new(id, name)
-    end
-  else
-    []
-  end
-rescue JSON::ParserError => e
-  puts "Error parsing 'genre.json': #{e.message}"
-  []
-end
-# load and save data to json file [END]......
-
+# load music genres   [START]>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
 genres = genres_reader.read
 music_albums = music_albums_reader.read
 puts 'No genres found in genre.json. You can add genres using the app.' if genres.empty?
@@ -301,5 +224,5 @@ loop do
   end
 end
 # Load genres and albums from JSON files
-genres = load_genres
-music_albums = load_music_albums(genres)
+genres = LoadMusic.load_genres
+music_albums = LoadMusic.load_music_albums(genres)
