+++ music_helper.rb
@@ -1,4 +1,26 @@
-# list methods
+require_relative 'write_music'
+require_relative 'load_music'
+
+# Save music albums to the JSON file
+def save_music_albums(music_albums)
+  WriteMusic.save_music_albums(music_albums)
+end
+
+# save genre
+def save_genres(genres)
+  WriteMusic.save_genres(genres)
+end
+
+# load music albums
+def load_music_albums(genres)
+  LoadMusic.load_music_albums(genres)
+end
+
+def find_or_create_genre(genres, genre_name)
+  LoadMusic.find_or_create_genre(genres, genre_name)
+end
+
+# list methods >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>
 def list_genres(genres)
   puts 'Listing all genres:'
   genres.each do |genre|
@@ -13,7 +35,7 @@ def list_all_music_albums(music_albums)
   end
 end
 
-# display methods
+# display methods music
 def display_music_album(index, album)
   puts "#{index}. Music Album -"
   display_music_album_details(album)
@@ -27,73 +49,4 @@ def display_music_album_details(album)
   puts "Genre: #{genre_name || 'No Genre'}"
 end
 
-# add music album
-# rubocop:disable Metrics/MethodLength
-def add_music_album(music_albums, items, genres)
-  puts 'Adding a new music album...'
-  print 'Enter album name: '
-  album_name = gets.chomp.to_s
-  print 'Enter genre name: '
-  genre_name = gets.chomp.to_s
-  print 'Is the album on Spotify? (true or false): '
-  on_spotify = gets.chomp.downcase == 'true'
-  print 'Can the album be archived? (true or false): '
-  can_be_archived = gets.chomp.downcase == 'true'
 
-  # rubocop:disable Lint/ShadowedArgument
-  genres = load_genres
-  # rubocop:enable Lint/ShadowedArgument
-
-  genre = genres.find { |g| g.name == genre_name }
-  unless genre
-    genre = Genre.new(genres.size + 1, genre_name)
-    genres << genre
-    save_genres(genres)
-  end
-
-  new_id = music_albums.size + 1
-  music_album = MusicAlbum.new(new_id, album_name, on_spotify, can_be_archived, genre)
-
-  music_album.can_be_archived = can_be_archived
-  items << music_album
-  music_albums << {
-    'id' => new_id,
-    'album' => {
-      'can_be_archived' => can_be_archived,
-      'on_spotify' => on_spotify,
-      'album_name' => album_name,
-      'genre_name' => genre_name
-    }
-  }
-
-  save_music_albums(music_albums)
-  # rubocop:enable Metrics/MethodLength
-  puts 'Music album added successfully.'
-end
-# save music
-# Save music albums to the JSON file
-
-def save_music_albums(music_albums)
-  data_to_save = music_albums.map do |album_data|
-    {
-      'id' => album_data['id'],
-      'album' => {
-        'can_be_archived' => album_data['album']['can_be_archived'],
-        'on_spotify' => album_data['album']['on_spotify'],
-        'album_name' => album_data['album']['album_name'],
-        # 'genre_id' => album_data['album']['genre_id']
-        'genre_name' => album_data['album']['genre_name']
-      }
-    }
-  end
-
-  json_file_path = 'music.json'
-
-  File.write(json_file_path, JSON.pretty_generate(data_to_save))
-
-  puts 'Music albums saved successfully.'
-rescue JSON::GeneratorError => e
-  puts "Error generating JSON data for '#{json_file_path}': #{e.message}"
-rescue StandardError => e
-  puts "Error saving music albums: #{e.message}"
-end
